Remove no longer needed tests
