+++ lib/trie/hash_test.go
@@ -4,9 +4,7 @@
 package trie
 
 import (
-	"bytes"
 	"errors"
-	"math/rand"
 	"testing"
 
 	"github.com/golang/mock/gomock"
@@ -18,68 +16,6 @@ type writeCall struct {
 	written []byte
 	n       int
 	err     error
-}
-
-func generateRandBytes(size int) []byte {
-	buf := make([]byte, rand.Intn(size)+1)
-	rand.Read(buf)
-	return buf
-}
-
-func generateRand(size int) [][]byte {
-	rt := make([][]byte, size)
-	for i := range rt {
-		buf := make([]byte, rand.Intn(379)+1)
-		rand.Read(buf)
-		rt[i] = buf
-	}
-	return rt
-}
-
-func TestHashLeaf(t *testing.T) {
-	n := &leaf{key: generateRandBytes(380), value: generateRandBytes(64)}
-
-	buffer := bytes.NewBuffer(nil)
-	const parallel = false
-	err := encodeNode(n, buffer, parallel)
-
-	if err != nil {
-		t.Errorf("did not hash leaf node: %s", err)
-	} else if buffer.Len() == 0 {
-		t.Errorf("did not hash leaf node: nil")
-	}
-}
-
-func TestHashBranch(t *testing.T) {
-	n := &branch{key: generateRandBytes(380), value: generateRandBytes(380)}
-	n.children[3] = &leaf{key: generateRandBytes(380), value: generateRandBytes(380)}
-
-	buffer := bytes.NewBuffer(nil)
-	const parallel = false
-	err := encodeNode(n, buffer, parallel)
-
-	if err != nil {
-		t.Errorf("did not hash branch node: %s", err)
-	} else if buffer.Len() == 0 {
-		t.Errorf("did not hash branch node: nil")
-	}
-}
-
-func TestHashShort(t *testing.T) {
-	n := &leaf{
-		key:   generateRandBytes(2),
-		value: generateRandBytes(3),
-	}
-
-	encodingBuffer := bytes.NewBuffer(nil)
-	const parallel = false
-	err := encodeNode(n, encodingBuffer, parallel)
-	require.NoError(t, err)
-
-	digestBuffer := bytes.NewBuffer(nil)
-	err = hashNode(n, digestBuffer)
-	require.NoError(t, err)
-	assert.Equal(t, encodingBuffer.Bytes(), digestBuffer.Bytes())
 }
 
 var errTest = errors.New("test error")
